+++ w2/stateMachine.cpp
@@ -19,7 +19,7 @@ void StateMachine::act(float dt, flecs::world &ecs, flecs::entity entity)
       if (transition.first->isAvailable(ecs, entity))
       {
         states[curStateIdx]->exit();
-        curStateIdx = transition.second;
+        curStateIdx = size_t(transition.second);
         states[curStateIdx]->enter();
         break;
       }
@@ -31,14 +31,14 @@ void StateMachine::act(float dt, flecs::world &ecs, flecs::entity entity)
 
 int StateMachine::addState(State *st)
 {
-  int idx = states.size();
+  size_t idx = states.size();
   states.push_back(st);
   transitions.push_back(std::vector<std::pair<StateTransition*, int>>());
-  return idx;
+  return int(idx);
 }
 
 void StateMachine::addTransition(StateTransition *trans, int from, int to)
 {
-  transitions[from].push_back(std::make_pair(trans, to));
+  transitions[size_t(from)].push_back(std::make_pair(trans, to));
 }
